Added Gulp to concatate and minify JS into a single file, source app.js is now stage.js, applied fix for leaking credentials, removed array syntax on controllers, added user database sync funtionality and user sort and filtering for admin page, updated add-user template, updated default background image

+++ server/controllers/db.js
@@ -24,16 +24,31 @@ exports.syncDb = function(req, res) {
     var getc = 'mongodump -h ' + rawCfg[getCfgKey].dbHost + (rawCfg[getCfgKey].dbUser ? ' -u ' + rawCfg[getCfgKey].dbUser : '') + (rawCfg[getCfgKey].dbPass ? ' -p ' + rawCfg[getCfgKey].dbPass : '') + ' -d ' + rawCfg[getCfgKey].db + ' -c users -o "' + __dirname + '\\..\\db"';
     var setc = 'mongorestore -h ' + rawCfg[setCfgKey].dbHost + (rawCfg[setCfgKey].dbUser ? ' -u ' + rawCfg[setCfgKey].dbUser : '') + (rawCfg[setCfgKey].dbPass ? ' -p ' + rawCfg[setCfgKey].dbPass : '') +' -d ' + rawCfg[setCfgKey].db + ' --drop "' + __dirname + '\\..\\db\\' + rawCfg[getCfgKey].db + '"';
     
+    var errenc = false;
     exec(getc, function(err, stdout, stderr) {
-      if(err) { console.log(err); }
-      console.log('Merging databases');
-      exec(setc, function(err, stdout, stderr) {
-        if(err) { console.log(err); }
-        console.log('Database sync has completed');
-        console.log('Removing temporary files');
-        rimraf(__dirname + '\\..\\db\\' + rawCfg[getCfgKey].db, function(err) { if(err) { console.log('Failed to remove temporary files: ' + err); } });
-        res.send({success: true});
-      })
+      if(err) { console.log(err); errenc = true; }
+      if(!errenc) {
+        console.log('Merging databases');
+        exec(setc, function(err, stdout, stderr) {
+          if(err) { console.log(err); errenc = true; }
+          if(!errenc) {
+            console.log('Database sync has completed');
+            console.log('Removing temporary files');
+            rimraf(__dirname + '\\..\\db\\' + rawCfg[getCfgKey].db, function(err) { if(err) { console.log('Failed to remove temporary files: ' + err); } });
+            res.send({success: true});
+          }
+          else {
+            console.log('Failed to restore database'); 
+            res.status(500); 
+            return res.end();
+          }
+        });
+      }
+      else { 
+        console.log('Failed to dump database'); 
+        res.status(500); 
+        return res.end();
+      }
     });
   }
